fix: fixed an error with submitting inline style in post

+++ donzoby.com/database/migrations/0001_01_01_000000_create_users_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('gender');
             $table->string('avatar');
             $table->string('country');
-            $table->integer('post_count')->default(0);
+            $table->unsignedBigInteger('post_count')->default(0);
             $table->enum('status', ['active', 'suspended'])->default('active');
             $table->string('password');
             $table->rememberToken();
